+++ src/roswire/ros2/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .ros2 import ROS2
+from .node_manager import ROS2NodeManager
+from .service_manager import ROS2ServiceManager